RemindersWebAPI/Reminders.BLL/CQS/Mediator.cs
using Microsoft.Extensions.DependencyInjection;
using Reminders.BLL.Interfaces;

namespace Reminders.BLL.CQS;

public class Mediator : IMediator
{
    private readonly IServiceProvider _serviceProvider;

    public Mediator(IServiceProvider serviceProvider)
    {
        _serviceProvider = serviceProvider;
    }

    public async Task SendCommandAsync<TCommand>(TCommand command)
    {
        using (var scope = _serviceProvider.CreateScope())
        {
            var handler = scope.ServiceProvider.GetRequiredService<ICommandHandler<TCommand>>();
            await handler.HandleAsync(command);
        }
    }

    public async Task<TResult> SendCommandAsync<TCommand, TResult>(TCommand command)
    {
        using (var scope = _serviceProvider.CreateScope())
        {
            var handler = scope.ServiceProvider.GetRequiredService<ICommandHandler<TCommand, TResult>>();
            return await handler.HandleAsync(command);
        }
    }
    
    public async Task<TResult> SendQueryAsync<TQuery, TResult>(TQuery query)
    {
        using (var scope = _serviceProvider.CreateScope())
        {
            var handler = scope.ServiceProvider.GetRequiredService<IQueryHandler<TQuery, TResult>>();
            return await handler.HandleAsync(query);
        }
    }
}